[Ruchir]Added: Lombork library

+++ src/main/java/com/bridgelabz/addressbook/service/IAddressBookService.java
@@ -13,6 +13,6 @@ public interface IAddressBookService {
 	AddressBookData addAddressBook(AddressBookDTO dto);
 	
 	AddressBookData updateAddressBook(int addressId,AddressBookDTO dto);
-	
-	void deleteAddressBook(long id);
+
+	void deleteAddressBook(int id);
 }
